+++ include/Qounters/NotesLeftQounter.hpp
@@ -14,7 +14,7 @@
 DECLARE_CLASS_CODEGEN(QountersMinus::Qounters, NotesLeftQounter, QountersMinus::Qounter,
     DECLARE_STATIC_FIELD(bool, Enabled);
     DECLARE_STATIC_FIELD(int, Position);
-    DECLARE_STATIC_FIELD(int, Distance);
+    DECLARE_STATIC_FIELD(float, Distance);
     DECLARE_STATIC_FIELD(bool, LabelAboveCount);
 
     DECLARE_INSTANCE_FIELD_DEFAULT(int, notesLeft, 0);
